# Conflicts:
#	app/src/main/java/com/app/smartpos/Registration/RegistrationDialog.java
#	app/src/main/java/com/app/smartpos/database/DatabaseOpenHelper.java

+++ app/src/main/java/com/app/smartpos/common/DeviceFactory/Device.java
@@ -4,7 +4,7 @@ import android.content.Intent;
 
 public class Device implements DeviceInterface{
     @Override
-    public Intent pay(double total) {
+    public Intent pay(long total) {
         return null;
     }
